+++ plugins/gradle-plugin-convention/src/main/kotlin/eu/bitfunk/gradle/plugin/convention/GradleExtensions.kt
@@ -1,0 +1,48 @@
+/*
+ * ISC License
+ *
+ * Copyright (c) 2022. Wolf-Martell Montwé (bitfunk)
+ *
+ * Permission to use, copy, modify, and/or distribute this software for any
+ * purpose with or without fee is hereby granted, provided that the above
+ * copyright notice and this permission notice appear in all copies.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS" AND THE AUTHOR DISCLAIMS ALL WARRANTIES WITH
+ * REGARD TO THIS SOFTWARE INCLUDING ALL IMPLIED WARRANTIES OF MERCHANTABILITY
+ * AND FITNESS. IN NO EVENT SHALL THE AUTHOR BE LIABLE FOR ANY SPECIAL, DIRECT,
+ * INDIRECT, OR CONSEQUENTIAL DAMAGES OR ANY DAMAGES WHATSOEVER RESULTING FROM
+ * LOSS OF USE, DATA OR PROFITS, WHETHER IN AN ACTION OF CONTRACT, NEGLIGENCE OR
+ * OTHER TORTIOUS ACTION, ARISING OUT OF OR IN CONNECTION WITH THE USE OR
+ * PERFORMANCE OF THIS SOFTWARE.
+ */
+
+package eu.bitfunk.gradle.plugin.convention
+
+import org.gradle.api.Action
+import org.gradle.api.Project
+import org.gradle.api.artifacts.Dependency
+import org.gradle.api.artifacts.dsl.DependencyHandler
+import org.gradle.api.plugins.JavaPluginExtension
+import org.gradle.testing.jacoco.plugins.JacocoPluginExtension
+
+public fun Project.javaPlugin(action: Action<JavaPluginExtension>) {
+    extensions.configure(JavaPluginExtension::class.java, action)
+}
+
+public fun Project.jacoco(action: Action<JacocoPluginExtension>) {
+    extensions.configure(JacocoPluginExtension::class.java, action)
+}
+
+public fun DependencyHandler.implementation(dependencyNotation: Any): Dependency? =
+    add("implementation", dependencyNotation)
+
+public fun DependencyHandler.testImplementation(dependencyNotation: Any): Dependency? =
+    add("testImplementation", dependencyNotation)
+
+public fun DependencyHandler.testRuntimeOnly(dependencyNotation: Any): Dependency? =
+    add("testRuntimeOnly", dependencyNotation)
+
+
+
+
+