+++ plugins/gradle-plugin-convention/src/main/kotlin/eu/bitfunk/gradle/plugin/convention/GradleKotlinExtensions.kt
@@ -16,34 +16,15 @@
  * PERFORMANCE OF THIS SOFTWARE.
  */
 
-plugins {
-    id("eu.bitfunk.gradle.plugin.convention")
+package eu.bitfunk.gradle.plugin.convention
+
+import org.gradle.api.Action
+import org.gradle.api.Project
+import org.jetbrains.kotlin.gradle.dsl.KotlinJvmProjectExtension
+
+public fun Project.kotlinJvm(action: Action<KotlinJvmProjectExtension>) {
+    extensions.configure(KotlinJvmProjectExtension::class.java, action)
 }
 
-group = "eu.bitfunk.gradle.version.catalog"
-version = "0.1.0"
 
-gradlePlugin {
-    plugins.create("gradlePluginVersionCatalog") {
-        id = "eu.bitfunk.gradle.version.catalog"
-        implementationClass = "eu.bitfunk.gradle.version.catalog.VersionCatalogAccessorPlugin"
-    }
-}
 
-dependencies {
-    implementation("com.squareup:kotlinpoet:1.10.2")
-    implementation("com.fasterxml.jackson.dataformat:jackson-dataformat-toml:2.13.0")
-}
-
-apiValidation {
-    ignoredPackages.add("eu.bitfunk.gradle.version.catalog.accessor")
-}
-
-tasks.register<Copy>("copySources") {
-    from("src/main/kotlin/eu/bitfunk/gradle/version/catalog/accessor")
-    into("src/main/resources/sources")
-}
-
-tasks.named("assemble") {
-    dependsOn("copySources")
-}