+++ CHANGELOG.md
@@ -17,5 +17,6 @@ Initial release
 
 ### Added
 
+* Add Gradle Plugin Convention
 * Add VersionCatalogAccessor - a version catalog accessor for Gradle Plugin development
 * Add initial project